# articles-service/docs/swagger.yaml
definitions:
  models.Article:
    properties:
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      photos:
        items:
          type: string
        maxItems: 3
        type: array
      timestamp:
        type: string
      title:
        maxLength: 140
        type: string
      user_id:
        type: integer
    required:
    - photos
    - title
    type: object
  models.Comment:
    properties:
      article_id:
        type: integer
      id:
        type: integer
      text:
        maxLength: 500
        type: string
      timestamp:
        type: string
      user_id:
        type: integer
    required:
    - article_id
    - text
    type: object
  models.User:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /article/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID объявления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Article'
        "400":
          description: Barticle Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получение одного объявления
      tags:
      - articles
  /article/all/:
    get:
      consumes:
      - application/json
      parameters:
      - description: Номер страницы
        in: query
        name: page
        required: true
        type: integer
      - description: Сортировка по дате (asc, desc)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Article'
            type: array
        "400":
          description: Barticle Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получение страницы Статей
      tags:
      - articles
  /article/create/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Объявление
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.Article'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: int
        "400":
          description: Barticle Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создание объявления
      tags:
      - articles
  /comments/{articleId}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получение всех комментариев статьи
      tags:
      - comments
  /comments/create/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Комментарий
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: int
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создание комментария
      tags:
      - comments
  /user/login/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Почта и пароль пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Аутентификация пользователя
      tags:
      - auth
  /user/register/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Почта и пароль пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
swagger: "2.0"